All toggle buttons function
For the toggle audio and toggle rain, an audio tag is added to html so that it changes according to what button is pressed.

The toggle sound button will add/remove the muted attribute and the toggle rain button will change the audio file to rain.wav/storm.wav.

The text on the page will change based on whether rain/storm is selected.